refactor: make updateDashboardIndexes be separate module
